+++ Team2Project/src/main/java/com/Team2Project/WorkWave/controller/AjaxController.java
@@ -24,7 +24,8 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 
 
-@RestController("/ajax")
+@RestController
+@RequestMapping("/ajax")
 public class AjaxController {
 
    @Autowired private ChatMapper chatMapper;
