+++ Team2Project/src/main/java/com/Team2Project/WorkWave/controller/LoginController.java
@@ -1,0 +1,14 @@
+package com.Team2Project.WorkWave.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+@Controller
+public class LoginController {
+
+	@GetMapping("/login")
+    public String userLogin() {
+       return "mainLogin";
+    }
+	
+}
