+++ Team2Project/src/main/resources/templates/include/header.html
@@ -12,7 +12,7 @@
             <div class="top-section">
                <div class="logo-search">
                   <div class="logo">
-                     <a th:href="@{/}">
+                     <a th:href="@{/A/main}">
                         <img src="/image/icon/main_logo.png" alt="Logo">
                      </a>
                   </div>
@@ -26,18 +26,18 @@
                </div>
                
                <div class="user-actions">
-                  <a th:href="@{/login.go}">로그인</a>
-                  <a th:href="@{/user_insert.go}">회원가입</a>
+                  <a th:href="@{/login}">로그인</a>
+                  <a th:href="@{/G/userJoin}">회원가입</a>
                </div>
             </div>
             
             <!-- 네비게이션 바를 검색창 아래로 이동 -->
             <nav>
                <ul>
-                  <li><a href="#">채용정보</a></li>
-                  <li><a th:href="@{/userFAQ.go}">고객센터</a></li>
-                  <li><a th:href="@{/notice.go}">공지사항</a></li>
-                  <li><a th:href="@{/chat.go}">웨이브톡</a></li>
+                  <li><a th:href="@{/A/comBoard}">채용정보</a></li>
+                  <li><a th:href="@{/A/userFAQ}">고객센터</a></li>
+                  <li><a th:href="@{/A/notice}">공지사항</a></li>
+                  <li><a th:href="@{/A/chat}">웨이브톡</a></li>
                </ul>
             </nav>
             
@@ -51,7 +51,7 @@
             <div class="top-section">
                <div class="logo-search">
                   <div class="logo">
-                     <a th:href="@{/}">
+                     <a th:href="@{/A/main}">
                         <img src="/image/icon/main_logo.png" alt="Logo">
                      </a>
                   </div>
@@ -76,10 +76,10 @@
             <!-- 네비게이션 바를 검색창 아래로 이동 -->
             <nav>
                <ul>
-                  <li><a href="#">채용정보</a></li>
-                  <li><a th:href="@{/userFAQ.go}">고객센터</a></li>
-                  <li><a th:href="@{/notice.go}">공지사항</a></li>
-                  <li><a th:href="@{/chat.go}">웨이브톡</a></li>
+                  <li><a th:href="@{/A/comBoard}">채용정보</a></li>
+                  <li><a th:href="@{/A/userFAQ}">고객센터</a></li>
+                  <li><a th:href="@{/A/notice}">공지사항</a></li>
+                  <li><a th:href="@{/A/chat}">웨이브톡</a></li>
                </ul>
             </nav>
             
@@ -93,7 +93,7 @@
             <div class="top-section">
                <div class="logo-search">
                   <div class="logo">
-                     <a th:href="@{/}">
+                     <a th:href="@{/A/main}">
                         <img src="/image/icon/main_logo.png" alt="Logo">
                      </a>
                   </div>
@@ -115,11 +115,11 @@
             <!-- 네비게이션 바를 검색창 아래로 이동 -->
             <nav>
                <ul>
-                  <li><a href="#">채용정보</a></li>
-                  <li><a th:href="@{/userFAQ.go}">고객센터</a></li>
-                  <li><a th:href="@{/notice.go}">공지사항</a></li>
-                  <li><a th:href="@{/chat.go}">웨이브톡</a></li>
-                  <li><a th:href="@{/comBoard/add}">공고등록</a></li>
+                  <li><a th:href="@{/A/comBoard}">채용정보</a></li>
+                  <li><a th:href="@{/A/userFAQ}">고객센터</a></li>
+                  <li><a th:href="@{/A/notice}">공지사항</a></li>
+                  <li><a th:href="@{/A/chat}">웨이브톡</a></li>
+                  <li><a th:href="@{/C/comBoard/insert}">공고등록</a></li>
                </ul>
             </nav>
