+++ Team2Project/src/main/resources/templates/index.html
@@ -10,18 +10,16 @@
 	<div th:replace="~{include/header}"></div>
 	<div layout:fragment="content" class="content">
 		<ul>
-			<li><a th:href="@{/login.go}">개인로그인</a></li>
-			<li><a th:href="@{/login.go}">기업로그인</a></li>
-			<li><a th:href="@{user_insert.go}">개인회원가입</a></li>
-			<li><a th:href="@{company_insert.go}">기업회원가입</a></li>
-			<li><a th:href="@{main.go}">비회원</a></li>
-			<li><a th:href="@{/comBoard}">공고리스트</a></li>
-			<li><a th:href="@{userFAQ.go}">고객센터</a></li>
-			<li><a th:href="@{profile_writer}">이력서 작성</a></li>
-			<li><a th:href="@{profile_list}">이력서 리스트</a></li>
-			<li><a th:href="@{/chat}">유저 게시판</a></li>
-			<li><a th:href="@{/comBoard/add}">공고등록</a></li>
-			<li><a th:href="@{/comBoard/update?No=18}">공고수정</a></li>
+			<li><a th:href="@{/login}">통합로그인</a></li>
+			<li><a th:href="@{/G/userJoin}">개인회원가입</a></li>
+			<li><a th:href="@{/G/companyJoin}">기업회원가입</a></li>
+			<li><a th:href="@{/A/comBoard}">공고리스트</a></li>
+			<li><a th:href="@{/A/userFAQ}">고객센터</a></li>
+			<li><a th:href="@{/U/profile/insert}">이력서 작성</a></li>
+			<li><a th:href="@{/U/profile}">이력서 리스트</a></li>
+			<li><a th:href="@{/A/chat}">유저 게시판</a></li>
+			<li><a th:href="@{/C/comBoard/insert}">공고등록</a></li>
+			<li><a th:href="@{/C/comBoard/update?No=18}">공고수정</a></li>
 		</ul>
 	</div>
 	<div th:replace="~{include/footer :: footer}"></div>
